aquacam_api/requirements.txt
flask
torch
pandas
Pillow
numpy